Create call & emit rewards
This code defines a RewardsDistributor contract that has a distributeRewards function that can be called by an iOS app to distribute rewards to users. The contract takes the address of a rewards contract as an input in its constructor. The distributeRewards function queries the rewards contract to get the list of users and their reward amounts, then calls the rewards contract's transfer function to send the rewards to the users. It also emits a RewardsDistributed event to track the distribution of rewards.